+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## 2.0.2 - 26/08/2026
+### Bugs
+- Fix a security issue that let a user with only read access to a group activate or deactivate other users' group memberships
+- Fix a redirect after activating/deactivating a group membership that could send the browser to an external site
+- Fix a security issue where the activate/deactivate group membership action could be triggered from another website without the user's intent
+
 ## 2.0.1 - 18/08/2026
 ### Features
 - Published an online user manual